factored out oidc utils
